requirements.txt
python-opencv
pyserial
numpy
opencv-python
opencv-python-headless